ルーティング処理の修正
