+++ kadai-4/src/pages/dashboard.vue
@@ -13,9 +13,6 @@
 /* eslint-disable */
   export default {
     async created(){
-      console.log('onAuth呼び出し')
-      await this.$store.dispatch('onAuth')
-      console.log('onAuthの処理終了')
       let loggedIn = this.$store.getters.loggedIn
       if (loggedIn === false){
         console.log('ログイン画面へ遷移')
