+++ kadai-4/src/pages/register.vue
@@ -23,7 +23,6 @@
 /* eslint-disable */
   export default {
     created() {
-      this.$store.dispatch('onAuth')
       if (this.$store.getters.loggedIn){
         this.$router.push({
           name: 'dashboard',
