+++ kadai-4/src/router/index.js
@@ -5,6 +5,7 @@ import register from '../pages/register'
 import login from '../pages/login'
 import dashboard from '../pages/dashboard'
 import firebase from 'firebase'
+import store from '../store/index'
 
 Vue.use(VueRouter)
 
@@ -18,17 +19,49 @@ const routes = [
       if (requiresAuth) {
         // このルートはログインされているかどうか認証が必要です。
         // もしされていないならば、ログインページにリダイレクトします。
-        firebase.auth().onAuthStateChanged(function (user) {
-          if (user) {
-            next({
-              path: '/dashboard',
-            })
-            // console.log('user',user)
-            // console.log(next)
-          } else {
-            next()
-          }
-        })
+        try {
+          firebase.auth().onAuthStateChanged(function (user) {
+            if (user) {
+              console.log('ログイン状態の取得')
+              user = user ? user : {}
+              let loginFlag = false
+              const db = firebase.firestore()
+              console.log('データベース情報取得')
+              if (user.uid){
+                commit('setUserUid', user.uid)
+                console.log('ユーザーIDの保存')
+                loginFlag = true
+                db
+                .collection('user-data')
+                .doc(user.uid)
+                .get()
+                .then(doc => {
+                  const userName = doc.data().userName
+                  const wallet = doc.data().wallet
+                  const payload = {
+                    userName: userName,
+                    wallet: wallet,
+                    loggedIn:true
+                  }
+                  commit('saveUserData', payload)
+                  console.log('ユーザー情報保存')
+                })
+                .catch(error => {
+                  console.log('エラー',error)
+                })
+              }
+              next({
+                path: '/dashboard',
+              })
+              // console.log('user',user)
+              // console.log(next)
+            } else {
+              next()
+            }
+          })
+        } catch (error) {
+          console.log(error)
+        }
       } else {
         next() // next() を常に呼び出すようにしてください!
       }
@@ -41,17 +74,49 @@ const routes = [
     beforeEnter: (to, from, next) => {
       const requiresAuth = to.matched.some(record => record.meta.requiresAuth)
       if (requiresAuth) {
-        firebase.auth().onAuthStateChanged(function (user) {
-          if (user) {
-            next({
-              path: '/dashboard',
-            })
-            console.log('user',user)
-            console.log(next)
-          } else {
-            next()
-          }
-        })
+        try {
+          firebase.auth().onAuthStateChanged(function (user) {
+            if (user) {
+              console.log('ログイン状態の取得')
+              user = user ? user : {}
+              let loginFlag = false
+              const db = firebase.firestore()
+              console.log('データベース情報取得')
+              if (user.uid){
+                store.commit('setUserUid', user.uid)
+                console.log('ユーザーIDの保存')
+                loginFlag = true
+                db
+                .collection('user-data')
+                .doc(user.uid)
+                .get()
+                .then(doc => {
+                  const userName = doc.data().userName
+                  const wallet = doc.data().wallet
+                  const payload = {
+                    userName: userName,
+                    wallet: wallet,
+                    loggedIn:true
+                  }
+                  store.commit('saveUserData', payload)
+                  console.log('ユーザー情報保存')
+                })
+                .catch(error => {
+                  console.log('エラー',error)
+                })
+              }
+              next({
+                path: '/dashboard',
+              })
+              // console.log('user',user)
+              // console.log(next)
+            } else {
+              next()
+            }
+          })
+        } catch (error) {
+          console.log(error)
+        }
       } else {
         next() // next() を常に呼び出すようにしてください!
       }
@@ -63,19 +128,53 @@ const routes = [
     props: true,
     meta: { requiresAuth: true },
     beforeEnter: (to, from, next) => {
+      console.log('to',to)
       const requiresAuth = to.matched.some(record => record.meta.requiresAuth)
       if (requiresAuth) {
-        firebase.auth().onAuthStateChanged(function (user) {
-          if (user) {
-            next()
-            // console.log('user',user)
-            // console.log(next)
-          } else {
-            next({
-              path: '/login',
-            })
-          }
-        })
+        try {
+          firebase.auth().onAuthStateChanged(function (user) {
+            if (user) {
+              console.log('ログイン状態の取得')
+              user = user ? user : {}
+              let loginFlag = false
+              const db = firebase.firestore()
+              console.log('データベース情報取得')
+              if (user.uid){
+                store.commit('setUserUid', user.uid)
+                console.log('ユーザーIDの保存')
+                loginFlag = true
+                db
+                .collection('user-data')
+                .doc(user.uid)
+                .get()
+                .then(doc => {
+                  const userName = doc.data().userName
+                  const wallet = doc.data().wallet
+                  const payload = {
+                    userName: userName,
+                    wallet: wallet,
+                    loggedIn:true
+                  }
+                  store.commit('saveUserData', payload)
+                  console.log('ユーザー情報保存')
+                })
+                .catch(error => {
+                  console.log('エラー',error)
+                })
+              }
+              next()
+              // console.log('user',user)
+              // console.log(next)
+            } else {
+              next({
+                path: '/login',
+              })
+            }
+          })
+        } catch (error) {
+          console.log(error)
+        }
+
       } else {
         next() // next() を常に呼び出すようにしてください!
       }
